requirements.txt
httprunner==1.5.8
djangorestframework
drf-yasg
django-cors-headers
django-filter
djangorestframework-jwt